ping h1 to h2 was throwing errors because the pop mpls function wasn't working on a 2 label stack - the ethertype variable was incorrectly being set to 0x800 instead of the mpls value. Setting the ethertype for mpls fixed this, and I confirmed with wireshark that 2 labels were attached by s1 on the way to s4, and 1 popped after s4. This then causes issues popping the last tag, as the current setuo doesn't have different rules for popping the last tag (which must go back to IP) or popping the 2nd tag (which must go to mpls).
About to implement a host specific mpls label which the edge node can pop off before sending directly to the host. This should hopefully fix it without adding any extra rules.

Also, there seems to be some issue with backup rules where an LFA is possible - just the output should be different, but h1 to h3 backup path adds an mpls tag for s4 before forwarding to s4 - which causes s4 to not know what to do.